requirements.txt
backports.zoneinfo==0.2.1
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///opt/concourse/worker/volumes/live/506ff722-306a-4142-6379-cfea99aadbed/volume/cffi_1625814701363/work
click @ file:///tmp/build/80754af9/click_1621604852318/work
clickhouse-driver==0.2.2
cryptography @ file:///opt/concourse/worker/volumes/live/c515855a-effc-46df-74dc-542901b701da/volume/cryptography_1616769282442/work
fastavro==1.4.4
Flask==1.0.2
grpcio==1.41.0
grpcio-tools==1.41.0
importlib-metadata @ file:///opt/concourse/worker/volumes/live/75f9b854-f895-4cdc-46a8-b054eddb9b3f/volume/importlib-metadata_1631916701668/work
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1624781299557/work
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
libsm3py==0.0.1
libsmdev-python==20210418
MarkupSafe @ file:///opt/concourse/worker/volumes/live/c9141381-1dba-485b-7c96-99007bf7bcfd/volume/markupsafe_1621528150226/work
mkl-fft==1.3.0
mkl-random @ file:///opt/concourse/worker/volumes/live/f196a661-8e33-4aac-63bc-efb2ff50e035/volume/mkl_random_1626186080069/work
mkl-service==2.4.0
numexpr @ file:///opt/concourse/worker/volumes/live/e845d683-bbb9-4fa2-79ce-743b84c61560/volume/numexpr_1618856522192/work
numpy @ file:///opt/concourse/worker/volumes/live/0afd8fc7-bc5d-4daa-6017-99ebdf04f26e/volume/numpy_and_numpy_base_1626271794342/work
pandas==1.3.0
peewee==3.14.4
protobuf==3.17.3
pulsar-client==2.8.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyMySQL @ file:///opt/concourse/worker/volumes/live/91bc5664-f34a-482d-7efc-f19bbd7a1347/volume/pymysql_1610482895985/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq @ file:///opt/concourse/worker/volumes/live/70168cdd-6dcc-4eea-5d88-71effee1f426/volume/pyzmq_1628276017199/work
scikit-learn @ file:///opt/concourse/worker/volumes/live/27ad05e5-2546-4218-4550-464724750acc/volume/scikit-learn_1621370399611/work
scipy @ file:///opt/concourse/worker/volumes/live/7d10d993-3825-404e-6e5d-9947c19e8c6d/volume/scipy_1618855951189/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1628668762525/work
tzdata==2021.5
tzlocal==4.0.2
Werkzeug==0.14.1
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
